requirements.txt
streamlit
easyocr
deepface
opencv-python-headless
pillow
insightface
onnxruntime
geopy
matplotlib